Added prompt for automatic crontab installation

+++ nq-install.sh
@@ -2,8 +2,8 @@
 #
 # NodeQuery Agent Installation Script
 #
-# @version		1.0.4
-# @date			2014-07-08
+# @version		1.0.5
+# @date			2014-07-18
 # @copyright	(c) 2014 http://nodequery.com
 #
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
@@ -38,35 +38,53 @@ fi
 # Check if crontab is installed
 if [ ! -n "$(command -v crontab)" ]
 then
+
+	# Confirm crontab installation
+	echo "|" && read -p "|   Crontab is required and could not be found. Do you want to install it? [Y/n] " input_variable_install
+
 	# Attempt to install crontab
-	if [ -n "$(command -v apt-get)" ]
+	if [ -z $input_variable_install ] || [ $input_variable_install == "Y" ] || [ $input_variable_install == "y" ]
 	then
-		echo -e "|\n|   Notice: Installing required package 'cron' via 'apt-get'"
-	    apt-get -y update
-	    apt-get -y install cron
-	    service cron start
-	elif [ -n "$(command -v yum)" ]
+		if [ -n "$(command -v apt-get)" ]
+		then
+			echo -e "|\n|   Notice: Installing required package 'cron' via 'apt-get'"
+		    apt-get -y update
+		    apt-get -y install cron
+		elif [ -n "$(command -v yum)" ]
+		then
+			echo -e "|\n|   Notice: Installing required package 'cronie' via 'yum'"
+		    yum -y install cronie
+		    
+		    if [ ! -n "$(command -v crontab)" ]
+		    then
+		    	echo -e "|\n|   Notice: Installing required package 'vixie-cron' via 'yum'"
+		    	yum -y install vixie-cron
+		    fi
+		elif [ -n "$(command -v pacman)" ]
+		then
+			echo -e "|\n|   Notice: Installing required package 'cronie' via 'pacman'"
+		    pacman -S --noconfirm cronie
+		fi
+	fi
+	
+	if [ ! -n "$(command -v crontab)" ]
 	then
-		echo -e "|\n|   Notice: Installing required package 'cronie' via 'yum'"
-	    yum -y install cronie
-	    chkconfig crond on
-	    service crond start
-	elif [ -n "$(command -v pacman)" ]
-	then
-		echo -e "|\n|   Notice: Installing required package 'cronie' via 'pacman'"
-	    pacman -S --noconfirm cronie
-	    systemctl start cronie
-	    systemctl enable cronie
-	else
 	    # Show error
 	    echo -e "|\n|   Error: Crontab is required and could not be installed\n|"
 	    exit 1
-	fi
-else
-	# Check if cron is running
-	if [ -z "$(ps -Al | grep cron | grep -v grep)" ]
+	fi	
+fi
+
+# Check if cron is running
+if [ -z "$(ps -Al | grep cron | grep -v grep)" ]
+then
+	
+	# Confirm cron service
+	echo "|" && read -p "|   Cron is available but not running. Do you want to start it? [Y/n] " input_variable_service
+
+	# Attempt to start cron
+	if [ -z $input_variable_service ] || [ $input_variable_service == "Y" ] || [ $input_variable_service == "y" ]
 	then
-		# Attempt to start cron
 		if [ -n "$(command -v apt-get)" ]
 		then
 			echo -e "|\n|   Notice: Starting 'cron' via 'service'"
@@ -78,17 +96,18 @@ else
 			service crond start
 		elif [ -n "$(command -v pacman)" ]
 		then
-			echo -e "|\n|   Notice: Starting 'cronie' via systemctl"
+			echo -e "|\n|   Notice: Starting 'cronie' via 'systemctl'"
 		    systemctl start cronie
 		    systemctl enable cronie
 		fi
-		
-		# Check if cron was started
-		if [ -z "$(ps -Al | grep cron | grep -v grep)" ]
-		then
-			# Show warning
-			echo -e "|\n|   Warning: Cron is available but could not be started\n|"
-		fi
+	fi
+	
+	# Check if cron was started
+	if [ -z "$(ps -Al | grep cron | grep -v grep)" ]
+	then
+		# Show error
+		echo -e "|\n|   Error: Cron is available but could not be started\n|"
+		exit 1
 	fi
 fi
 
@@ -96,7 +115,7 @@ fi
 if [ -f /etc/nodequery/nq-agent.sh ]
 then
 	# Remove agent dir
-	rm -R /etc/nodequery
+	rm -Rf /etc/nodequery
 
 	# Remove cron entry and user
 	if id -u nodequery >/dev/null 2>&1
